requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
dnspython==1.16.0
dominate==2.4.0
eventlet==0.25.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==4.2.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
git+https://github.com/schmidtbri/ml-model-abc-improvements#egg=iris_model
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
marshmallow==3.5.0
monotonic==1.5
numpy==1.18.1
python-engineio==3.11.2
python-socketio==4.4.0
requests==2.23.0
schema==0.7.0
scikit-learn==0.21.0
scipy==1.4.1
six==1.14.0
urllib3==1.25.8
visitor==0.1.3
websocket-client==0.57.0
Werkzeug==1.0.0



